split files mongodb model
